Update dev branch

+++ listen-for-shutdown.py
@@ -29,14 +29,16 @@ while (True):
 		GPIO.setup(powerPin, GPIO.OUT)
 		GPIO.output(powerPin, 0);		
 		subprocess.call(['reboot', '-h', 'now'], shell=False)
+		release = True;
+		time.sleep(10);
 	GPIO.output(powerPin, 0); # blink once
 	time.sleep(0.1);
 	GPIO.output(powerPin, 1);
 	time.sleep(1)
-	if GPIO.input(26):
+	if (GPIO.input(26) and (release == False)):
 		print("switch to AFSK")
 		f = open("/home/pi/CubeSatSim/.mode", "w")
-		f.write("ARG1=a")
+		f.write("a")
 		f.close()
 		os.system("sudo systemctl stop cubesatsim")
 		os.system("sudo systemctl start cubesatsim")
@@ -53,7 +55,7 @@ while (True):
 	if (GPIO.input(26) and (release == False)):
 		print("switch to FSK")
 		f = open("/home/pi/CubeSatSim/.mode", "w")
-		f.write("ARG1=f")
+		f.write("f")
 		f.close()
 		os.system("sudo systemctl stop cubesatsim")
 		os.system("sudo systemctl start cubesatsim")
@@ -74,7 +76,7 @@ while (True):
 	if (GPIO.input(26) and (release == False)):
 		print("switch to BPSK")
 		f = open("/home/pi/CubeSatSim/.mode", "w")
-		f.write("ARG1=b")
+		f.write("b")
 		f.close()
 		os.system("sudo systemctl stop cubesatsim")
 		os.system("sudo systemctl start cubesatsim")
@@ -99,7 +101,7 @@ while (True):
 	if (GPIO.input(26) and (release == False)):
 		print("switch to SSTV")
 		f = open("/home/pi/CubeSatSim/.mode", "w")
-		f.write("ARG1=s")
+		f.write("s")
 		f.close()
 		os.system("sudo systemctl stop cubesatsim")
 		os.system("sudo systemctl start cubesatsim")
